feat: add `NodeTelemetry`, which implements both `PublisherTelemetry` and `ChainTelemetry`
